Grain Encapsulate a grid, with dimension and extent.

A grid is a raster, a domain with an underlying regular discretization in xy.

Public fields

dimension

dimension (nx, ny) of grid, number of columns, number of rows, defaults to 1x1 (unless 'world = TRUE')

extent

extent of grid xmin,xmax,ymin,ymax - defaults to 0,ncol,0,nrow

Methods

Public methods

grain$new()

grain$align_extent()

grain$cell_from_col()

grain$cell_from_extent()

grain$cell_from_row()

grain$cell_from_col_row()

grain$cell_from_colrow_combine()

grain$cell_from_xy()

grain$col_from_cell()

grain$col_from_x()

grain$colrow_from_cell()

grain$extent_dimension()

grain$extent_from_cell()

grain$extent_vrt()

grain$intersect_extent()

grain$listxyz()

grain$long()

grain$n_cell()

grain$n_col()

grain$n_row()

grain$origin()

grain$plot_extent()

grain$row_from_cell()

grain$row_from_y()

grain$snap_extent()

grain$vcrop()

grain$x_centre()

grain$x_corner()

grain$x_from_cell()

grain$x_from_col()

grain$x_max()

grain$x_min()

grain$x_res()

grain$xlim()

grain$xy()

grain$xy_from_cell()

grain$y_centre()

grain$y_corner()

grain$y_from_cell()

grain$y_from_row()

grain$y_max()

grain$y_min()

grain$y_res()

grain$ylim()

grain$greet()

grain$clone()

Method new()

Usage

grain$new(dimension = NULL, extent = NULL, world = FALSE)

Arguments

dimension

ncol, nrow

extent

xmin,xmax,ymin,ymax

world

create a world longlat grid with 1 degree cells

Examples

grain$new(c(360, 180), c(-180, 180, -90, 90))
 grain$new(c(360, 180), c(-180, 180, -90, 90))$align_extent(c(-179.5, 179, 0.03, 10))
g <- grain$new(world = TRUE)
g$cell_from_col(2)

Method align_extent()

Usage

grain$align_extent(x, snap = "out")

Arguments

x

x coordinate

snap

"out", "in", "near"

Returns

aligned extent

Method cell_from_col()

Usage

grain$cell_from_col(col)

Arguments

col

column index

Returns

cell index, extent, or coordinate

Method cell_from_extent()

Usage

grain$cell_from_extent(extent)

Arguments

extent

xmin,xmax,ymin,ymax

Returns

cell index

Examples

g <- grain$new(world = TRUE)
g$cell_from_extent(c(147, 148, -43, -41))

Method cell_from_row()

Usage

grain$cell_from_row(row)

Arguments

row

row index

Returns

cell index

Method cell_from_col_row()

Usage

grain$cell_from_col_row(col, row)

Arguments

col

column index

row

row index

Returns

cell index

Method cell_from_colrow_combine()

Usage

grain$cell_from_colrow_combine(col, row)

Arguments

col

column index

row

row index

Returns

cell index

Method cell_from_xy()

Usage

grain$cell_from_xy(xy)

Arguments

xy

xy coordinates

Returns

cell index

Method col_from_cell()

Usage

grain$col_from_cell(cell)

Arguments

cell

cell index

Returns

cell index

Method col_from_x()

Usage

grain$col_from_x(x)

Arguments

x

x coordinate

Returns

col index

Method colrow_from_cell()

Usage

grain$colrow_from_cell(cell)

Arguments

cell

cell index

Returns

col,row index

Method extent_dimension()

Usage

grain$extent_dimension(extent, snap = "out")

Arguments

extent

xmin,xmax,ymin,ymax

snap

"out", "in", "near"

Returns

dimension

Method extent_from_cell()

Usage

grain$extent_from_cell(cell)

Arguments

cell

cell index

Returns

extent

Method extent_vrt()

Usage

grain$extent_vrt(vrt)

Arguments

vrt

vrt file path

Returns

extents, in a matrix, xmin,xmax,ymin,ymax

Method intersect_extent()

Usage

grain$intersect_extent(extent)

Arguments

extent

xmin,xmax,ymin,ymax

Returns

extent

Method listxyz()

Usage

grain$listxyz(data = NULL)

Arguments

data

optional data values for output

Returns

list of x,y vectors and z matrix (like graphics::image())

Method long()

Usage

grain$long(data = NULL, raster_order = TRUE)

Arguments

data

optional data values for output

raster_order

raster order or traditional R matrix order

Returns

data frame of x, y and optionaly z value

Method n_cell()

Usage

grain$n_cell()

Returns

integer, number of cells

Method n_col()

Usage

grain$n_col()

Returns

integer, number of columns

Method n_row()

Usage

grain$n_row()

Returns

integer, number of rows

Method origin()

Usage

grain$origin()

Returns

coordinate of grid origin

Method plot_extent()

Usage

grain$plot_extent(..., extent = NULL)

Arguments

...

args passed to graphics::plot()

extent

xmin,xmax,ymin,ymax

Returns

nothing, called for side effect of plot

Method row_from_cell()

Usage

grain$row_from_cell(cell)

Arguments

cell

cell index

Returns

integer of row index

Method row_from_y()

Usage

grain$row_from_y(y)

Arguments

y

y coordinate

Returns

integer of row index

Method snap_extent()

Usage

grain$snap_extent(res)

Arguments

res

resolution

Returns

extent

Method vcrop()

Usage

grain$vcrop(extent, snap = "out")

Arguments

extent

xmin,xmax,ymin,ymax

snap

"out", "in", "near"

Returns

extent, dimension cropped (and snapped to grain)

Method x_centre()

Usage

grain$x_centre()

Returns

x coordinates of centre of cell

Method x_corner()

Usage

grain$x_corner()

Returns

x coordinates of corners (edges) of cell

Method x_from_cell()

Usage

grain$x_from_cell(cell)

Arguments

cell

cell index

Returns

x coordinate of cell (centre)

Method x_from_col()

Usage

grain$x_from_col(col)

Arguments

col

column index

Returns

x of column (centre)

Method x_max()

Usage

grain$x_max()

Returns

maximum x (right of grid)

Method x_min()

Usage

grain$x_min()

Returns

minimum x (left of grid)

Method x_res()

Usage

grain$x_res()

Returns

x resolution (width of cell)

Method xlim()

Usage

grain$xlim()

Returns

x range of grid (leftmost edge to rightmost edge)

Method xy()

Usage

grain$xy()

Returns

coordinates of cells (centre)

Method xy_from_cell()

Usage

grain$xy_from_cell(cell)

Arguments

cell

cell index

Examples

g <- grain$new(world = TRUE)
g$xy_from_cell(g$cell_from_extent(c(147, 148, -43, -41)))

Method y_centre()

Usage

grain$y_centre()

Returns

y coordinate (centre)

Method y_corner()

Usage

grain$y_corner()

Returns

y coordinates (corners)

Method y_from_cell()

Usage

grain$y_from_cell(cell)

Arguments

cell

cell index

Returns

y coordinate of cell (centre)

Method y_from_row()

Usage

grain$y_from_row(row)

Arguments

row

row index

Returns

y coordinate of row (centre)

Method y_max()

Usage

grain$y_max()

Returns

maximum y coordinate of grid (top edge)

Method y_min()

Usage

grain$y_min()

Returns

minimum y coordinate of grid (left edge)

Method y_res()

Usage

grain$y_res()

Returns

resolution y of grid (height of cell)

Method ylim()

Usage

grain$ylim()

Returns

extent of grid in y (bottom most edge to topmost edge)

Method greet()

Usage

grain$greet()

Returns

hello

Method clone()

The objects of this class are cloneable with this method.

Usage

grain$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `grain$new`
## ------------------------------------------------

grain$new(c(360, 180), c(-180, 180, -90, 90))
#> grain object
#> Dimension:  360x180
#>    Extent:  -180, 180, -90, 90 (xmin,xmax,ymin,ymax)
#> <grain>
#>   Public:
#>     align_extent: function (x, snap = "out") 
#>     cell_from_col: function (col) 
#>     cell_from_col_row: function (col, row) 
#>     cell_from_colrow_combine: function (col, row) 
#>     cell_from_extent: function (extent) 
#>     cell_from_row: function (row) 
#>     cell_from_xy: function (xy) 
#>     clone: function (deep = FALSE) 
#>     col_from_cell: function (cell) 
#>     col_from_x: function (x) 
#>     colrow_from_cell: function (cell) 
#>     dimension: 360 180
#>     extent: -180 180 -90 90
#>     extent_dimension: function (extent, snap = "out") 
#>     extent_from_cell: function (cell) 
#>     extent_vrt: function (vrt) 
#>     greet: function () 
#>     initialize: function (dimension = NULL, extent = NULL, world = FALSE) 
#>     intersect_extent: function (extent) 
#>     listxyz: function (data = NULL) 
#>     long: function (data = NULL, raster_order = TRUE) 
#>     n_cell: function () 
#>     n_col: function () 
#>     n_row: function () 
#>     origin: function () 
#>     plot_extent: function (..., extent = NULL) 
#>     row_from_cell: function (cell) 
#>     row_from_y: function (y) 
#>     snap_extent: function (res) 
#>     vcrop: function (extent, snap = "out") 
#>     x_centre: function () 
#>     x_corner: function () 
#>     x_from_cell: function (cell) 
#>     x_from_col: function (col) 
#>     x_max: function () 
#>     x_min: function () 
#>     x_res: function () 
#>     xlim: function () 
#>     xy: function () 
#>     xy_from_cell: function (cell) 
#>     y_centre: function () 
#>     y_corner: function () 
#>     y_from_cell: function (cell) 
#>     y_from_row: function (row) 
#>     y_max: function () 
#>     y_min: function () 
#>     y_res: function () 
#>     ylim: function () 
 grain$new(c(360, 180), c(-180, 180, -90, 90))$align_extent(c(-179.5, 179, 0.03, 10))
#> grain object
#> Dimension:  360x180
#>    Extent:  -180, 180, -90, 90 (xmin,xmax,ymin,ymax)
#> [1] -180  179    0   10
g <- grain$new(world = TRUE)
#> grain object
#> Dimension:  360x180
#>    Extent:  -180, 180, -90, 90 (xmin,xmax,ymin,ymax)
g$cell_from_col(2)
#>   [1]     2   362   722  1082  1442  1802  2162  2522  2882  3242  3602  3962
#>  [13]  4322  4682  5042  5402  5762  6122  6482  6842  7202  7562  7922  8282
#>  [25]  8642  9002  9362  9722 10082 10442 10802 11162 11522 11882 12242 12602
#>  [37] 12962 13322 13682 14042 14402 14762 15122 15482 15842 16202 16562 16922
#>  [49] 17282 17642 18002 18362 18722 19082 19442 19802 20162 20522 20882 21242
#>  [61] 21602 21962 22322 22682 23042 23402 23762 24122 24482 24842 25202 25562
#>  [73] 25922 26282 26642 27002 27362 27722 28082 28442 28802 29162 29522 29882
#>  [85] 30242 30602 30962 31322 31682 32042 32402 32762 33122 33482 33842 34202
#>  [97] 34562 34922 35282 35642 36002 36362 36722 37082 37442 37802 38162 38522
#> [109] 38882 39242 39602 39962 40322 40682 41042 41402 41762 42122 42482 42842
#> [121] 43202 43562 43922 44282 44642 45002 45362 45722 46082 46442 46802 47162
#> [133] 47522 47882 48242 48602 48962 49322 49682 50042 50402 50762 51122 51482
#> [145] 51842 52202 52562 52922 53282 53642 54002 54362 54722 55082 55442 55802
#> [157] 56162 56522 56882 57242 57602 57962 58322 58682 59042 59402 59762 60122
#> [169] 60482 60842 61202 61562 61922 62282 62642 63002 63362 63722 64082 64442

## ------------------------------------------------
## Method `grain$cell_from_extent`
## ------------------------------------------------

g <- grain$new(world = TRUE)
#> grain object
#> Dimension:  360x180
#>    Extent:  -180, 180, -90, 90 (xmin,xmax,ymin,ymax)
g$cell_from_extent(c(147, 148, -43, -41))
#> [1] 47488 47848

## ------------------------------------------------
## Method `grain$xy_from_cell`
## ------------------------------------------------

g <- grain$new(world = TRUE)
#> grain object
#> Dimension:  360x180
#>    Extent:  -180, 180, -90, 90 (xmin,xmax,ymin,ymax)
g$xy_from_cell(g$cell_from_extent(c(147, 148, -43, -41)))
#>       [,1]  [,2]
#> [1,] 147.5 -41.5
#> [2,] 147.5 -42.5